+++ imports/components/Input/Percent/index.js
@@ -1,0 +1,73 @@
+import React from 'react';
+import { Meteor } from 'meteor/meteor';
+
+import tools from '/imports/startup/tools/index';
+
+// This is the best use of Input. Other types should follow this.
+// The displayValue is only set after update in props.
+// When onChange happens, the component will only export the value
+// The update on display will happen after, in componentDidUpdate
+export default class Percent extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      displayValue: 0
+    }
+  }
+
+  componentDidMount() {
+    this.setDisplayValue(this.props.value);
+  }
+
+  componentDidUpdate(prevProps, prevState) {
+    if (prevProps.value !== this.props.value) {
+      this.setDisplayValue(this.props.value);
+    }
+  }
+
+  getDisplayValue = (displayValue) => {
+    function clamp(value) {
+      value = Number(value);
+
+      if (value > 100) return 100;
+      if (value < 0) return 0;
+      return value;
+    }
+    displayValue = tools.format(displayValue, 'number');
+    return clamp(displayValue);
+  }
+
+  setDisplayValue = (displayValue) => {
+    var displayValue = this.getDisplayValue(displayValue);
+    displayValue = Math.round(displayValue * 100);
+    this.setState({
+      displayValue: displayValue.toString()
+    });
+  }
+
+  onChange = (e) => {
+    e.preventDefault();
+    var displayValue = this.getDisplayValue(e.target.value);
+    var exportValue = displayValue / 100;
+    this.props.onChange(exportValue);
+  }
+
+  render() {
+    return (
+      <input
+        type="number"
+        value={this.state.displayValue}
+        onChange={this.onChange}
+
+        min={0}
+        max={100}
+
+        readOnly={this.props.readOnly}
+        placeholder={this.props.placeholder}
+        disabled={this.props.disabled}
+
+        style={this.props.style}
+        />
+    )
+  }
+}