+++ imports/scenes/Contract/Information/index.js
@@ -130,10 +130,7 @@ export default class Information extends React.Component {
             <Input
               title="Desconto:"
               name="discount"
-              type="number"
-              percent={true}
-              max={100}
-              min={0}
+              type="percent"
               disabled={this.props.disabled}
               error={this.props.errorKeys.includes("discount")}
               value={this.props.contract.discount}